Do not expose constructor locating as core functionality.
Since we have to specialize this for JUnit 3 and JUnit 4 there's no point in having it in the core.
